auth_microservice/pyproject.toml
[project]
name = "auth-microservice"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcus",email = "marc303i@stud.kea.dk"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
